Revert "Slack 메시지 전송 기능 구현 (#4)"
This reverts commit cf0c1fbb045d923c83a872aa7983208dea28a2f4.
