+++ .gitignore
@@ -38,6 +38,3 @@ out/
 
 ### Kotlin ###
 .kotlin
-
-### Environment Variable ###
-application.yml
