+++ build.gradle.kts
@@ -3,7 +3,7 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 plugins {
 	val kotlinVersion = "1.9.23"
 
-	id("org.springframework.boot") version "3.2.4" // Spring Ver: 6.1.5
+	id("org.springframework.boot") version "3.2.4"
 	id("io.spring.dependency-management") version "1.1.4"
 	kotlin("jvm") version kotlinVersion
 	kotlin("plugin.spring") version kotlinVersion
@@ -22,29 +22,19 @@ repositories {
 }
 
 dependencies {
-	// Spring
-	implementation("org.springframework.boot:spring-boot-starter-web")
-	implementation("org.springframework.boot:spring-boot-starter-data-jpa")
-	implementation("org.springframework.boot:spring-boot-starter-security") // Ver: 6.2.3
-	testImplementation("org.springframework.boot:spring-boot-starter-test")
-	testImplementation("org.springframework.security:spring-security-test")
-
-	// Kotlin
-	implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
-	implementation("org.jetbrains.kotlin:kotlin-reflect")
-
 	// DB
 	runtimeOnly("org.mariadb.jdbc:mariadb-java-client")
 
-	// JWT
-	implementation("io.jsonwebtoken:jjwt:0.12.5")
+	implementation("com.fasterxml.jackson.module:jackson-module-kotlin")
+	implementation("org.springframework.boot:spring-boot-starter-web")
+	implementation("org.springframework.boot:spring-boot-starter-data-jpa")
+	implementation("org.springframework.boot:spring-boot-starter-security")
+	implementation("org.jetbrains.kotlin:kotlin-reflect")
+	testImplementation("org.springframework.boot:spring-boot-starter-test")
+	testImplementation("org.springframework.security:spring-security-test")
 
 	// Logging
 	implementation("io.github.microutils:kotlin-logging-jvm:3.0.5")
-
-	// Slack
-	implementation("com.slack.api:slack-api-model-kotlin-extension:1.40.2")
-	implementation("com.slack.api:slack-api-client-kotlin-extension:1.40.2")
 }
 
 tasks.withType<KotlinCompile> {
