+++ src/main/kotlin/com/ohforbidden/bugreport/controller/Test.kt
@@ -1,0 +1,14 @@
+package com.ohforbidden.bugreport.controller
+
+import org.springframework.web.bind.annotation.GetMapping
+import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RestController
+
+@RestController
+@RequestMapping("/test/filter")
+class Test {
+    @GetMapping
+    fun filterTestTrigger() {
+        println("### Controller Works!! ###")
+    }
+}
