+++ src/main/kotlin/com/ohforbidden/bugreport/domain/user/UserRepository.kt
@@ -1,0 +1,5 @@
+package com.ohforbidden.bugreport.domain.user
+
+import org.springframework.data.jpa.repository.JpaRepository
+
+interface UserRepository : JpaRepository<User, String>
