+++ src/main/kotlin/com/ohforbidden/bugreport/global/dto/ErrorResponse.kt
@@ -1,0 +1,12 @@
+package com.ohforbidden.bugreport.global.dto
+
+import com.ohforbidden.bugreport.global.util.createUtcDateTime
+import java.time.ZonedDateTime
+
+data class ErrorResponse(
+    val uuid: String,
+    val httpStatus: Int,
+    val errorCode: String,
+    val message: String?,
+    val timeStamp: ZonedDateTime = createUtcDateTime()
+)
