+++ src/main/kotlin/com/ohforbidden/bugreport/global/exception/BusinessException.kt
@@ -1,0 +1,6 @@
+package com.ohforbidden.bugreport.global.exception
+
+class BusinessException(
+    val errorType: ErrorType,
+    override val cause: Throwable? = null,
+) : RuntimeException(errorType.message, cause)
