+++ src/main/kotlin/com/ohforbidden/bugreport/global/exception/ErrorType.kt
@@ -1,0 +1,9 @@
+package com.ohforbidden.bugreport.global.exception
+
+import org.springframework.http.HttpStatus
+
+interface ErrorType {
+    val httpStatus: HttpStatus
+    val errorCode: String
+    val message: String
+}
