+++ src/main/kotlin/com/ohforbidden/bugreport/global/util/TimeUtils.kt
@@ -1,0 +1,6 @@
+package com.ohforbidden.bugreport.global.util
+
+import java.time.ZoneOffset
+import java.time.ZonedDateTime
+
+fun createUtcDateTime() = ZonedDateTime.now(ZoneOffset.UTC)
