+++ src/main/kotlin/com/ohforbidden/bugreport/infrastructure/config/JpaConfig.kt
@@ -1,0 +1,8 @@
+package com.ohforbidden.bugreport.infrastructure.config
+
+import org.springframework.context.annotation.Configuration
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing
+
+@Configuration
+@EnableJpaAuditing
+class JpaConfig
