--- src/main/kotlin/com/ohforbidden/domain/user/UserRepository.kt
@@ -1,5 +1,0 @@
-package com.ohforbidden.domain.user
-
-import org.springframework.data.jpa.repository.JpaRepository
-
-interface UserRepository : JpaRepository<User, String>
