--- src/main/kotlin/com/ohforbidden/global/auth/JwtClaims.kt
@@ -1,7 +1,0 @@
-package com.ohforbidden.global.auth
-
-data class JwtClaims(
-    val userId: Long,
-    val email: String,
-    val role: Role
-)
