--- src/main/kotlin/com/ohforbidden/global/auth/LoginRequest.kt
@@ -1,11 +1,0 @@
-package com.ohforbidden.global.auth
-
-import com.fasterxml.jackson.annotation.JsonProperty
-
-data class LoginRequest(
-    @JsonProperty("email")
-    val email: String,
-
-    @JsonProperty("password")
-    val password: String
-)
