--- src/main/kotlin/com/ohforbidden/global/auth/LoginResponse.kt
@@ -1,6 +1,0 @@
-package com.ohforbidden.global.auth
-
-data class LoginResponse(
-    val accessToken: String,
-    val refreshToken: String
-)
