--- src/main/kotlin/com/ohforbidden/global/auth/TokenType.kt
@@ -1,6 +1,0 @@
-package com.ohforbidden.global.auth
-
-enum class TokenType(val subject: String) {
-    ACCESS("access_token"),
-    REFRESH("refresh_token")
-}
