--- src/main/kotlin/com/ohforbidden/global/config/EncoderConfig.kt
@@ -1,12 +1,0 @@
-package com.ohforbidden.global.config
-
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder
-
-@Configuration
-class EncoderConfig {
-
-    @Bean
-    fun encoder(): BCryptPasswordEncoder = BCryptPasswordEncoder()
-}
