--- src/main/kotlin/com/ohforbidden/global/config/WebConfig.kt
@@ -1,28 +1,0 @@
-package com.ohforbidden.global.config
-
-import com.ohforbidden.global.filter.MDCLoggingFilter
-import jakarta.servlet.Filter
-import org.springframework.boot.web.servlet.FilterRegistrationBean
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
-
-@Configuration
-class WebConfig() : WebMvcConfigurer {
-//    override fun addInterceptors(registry: InterceptorRegistry) {
-//        registry
-//            .addInterceptor(LogInterceptor())
-//            .order(1)
-//            .addPathPatterns("/**")
-//            .excludePathPatterns("/css/**", "/*.ico", "/error")
-//    }
-
-    @Bean
-    fun mdcLoggingFilter(): FilterRegistrationBean<Filter> {
-        val filterRegistrationBean = FilterRegistrationBean<Filter>()
-        filterRegistrationBean.filter = MDCLoggingFilter()
-        filterRegistrationBean.order = 1
-        filterRegistrationBean.addUrlPatterns("/*")
-        return filterRegistrationBean
-    }
-}
