--- src/main/kotlin/com/ohforbidden/global/exception/AuthException.kt
@@ -1,6 +1,0 @@
-package com.ohforbidden.global.exception
-
-class AuthException(
-    val errorType: ErrorType,
-    override val cause: Throwable? = null,
-) : RuntimeException(errorType.message, cause)
