--- src/main/kotlin/com/ohforbidden/global/exception/BusinessException.kt
@@ -1,6 +1,0 @@
-package com.ohforbidden.global.exception
-
-class BusinessException(
-    val errorType: ErrorType,
-    override val cause: Throwable? = null,
-) : RuntimeException(errorType.message, cause)
