--- src/main/kotlin/com/ohforbidden/global/exception/ErrorType.kt
@@ -1,9 +1,0 @@
-package com.ohforbidden.global.exception
-
-import org.springframework.http.HttpStatus
-
-interface ErrorType {
-    val httpStatus: HttpStatus
-    val errorCode: String
-    val message: String
-}
