--- src/main/kotlin/com/ohforbidden/global/util/TimeUtils.kt
@@ -1,6 +1,0 @@
-package com.ohforbidden.global.util
-
-import java.time.ZoneOffset
-import java.time.ZonedDateTime
-
-fun createUtcDateTime() = ZonedDateTime.now(ZoneOffset.UTC)
