--- src/main/kotlin/com/ohforbidden/infrastructure/config/JpaConfig.kt
@@ -1,8 +1,0 @@
-package com.ohforbidden.infrastructure.config
-
-import org.springframework.context.annotation.Configuration
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing
-
-@Configuration
-@EnableJpaAuditing
-class JpaConfig
