--- src/main/kotlin/com/ohforbidden/infrastructure/message/slack/config/SlackConfig.kt
@@ -1,11 +1,0 @@
-package com.ohforbidden.infrastructure.message.slack.config
-
-import com.slack.api.Slack
-import org.springframework.context.annotation.Bean
-import org.springframework.context.annotation.Configuration
-
-@Configuration
-class SlackConfig {
-    @Bean
-    fun slack(): Slack = Slack.getInstance()
-}
