--- src/main/kotlin/com/ohforbidden/infrastructure/message/slack/dto/SignUpParam.kt
@@ -1,7 +1,0 @@
-package com.ohforbidden.infrastructure.message.slack.dto
-
-data class SignUpParam(
-    val username: String,
-    val email: String,
-    val code: String
-)
