+++ src/main/resources/application.yml
@@ -1,0 +1,18 @@
+spring:
+    # MariaDB
+    datasource:
+        driver-class-name: org.mariadb.jdbc.Driver
+        url: jdbc:mariadb://localhost:3306/bugreport
+        username: admin
+        password: 1234
+
+    # Table Auto Create Test
+    jpa:
+        hibernate:
+          ddl-auto: none
+
+    # User
+    security:
+        user:
+            name: user
+            password: 1234
